+++ apps/ehr/src/features/visits/shared/components/assessment-tab/EMCodeField.tsx
@@ -2,12 +2,8 @@ import { Autocomplete, TextField } from '@mui/material';
 import { enqueueSnackbar } from 'notistack';
 import { FC } from 'react';
 import { dataTestIds } from 'src/constants/data-test-ids';
+import { CPTCodeOption, emCodeOptions } from 'utils';
 import { useChartData, useDeleteChartData, useSaveChartData } from '../../stores/appointment/appointment.store';
-
-export type CPTCodeOption = {
-  code: string;
-  display: string;
-};
 
 export const EMCodeField: FC = () => {
   const { chartData, setPartialChartData } = useChartData();
@@ -55,15 +51,3 @@ export const EMCodeField: FC = () => {
     />
   );
 };
-
-export const emCodeOptions: CPTCodeOption[] = [
-  { display: '99201 New Patient - E/M Level 1', code: '99201' },
-  { display: '99202 New Patient - E/M Level 2', code: '99202' },
-  { display: '99203 New Patient - E/M Level 3', code: '99203' },
-  { display: '99204 New Patient - E/M Level 4', code: '99204' },
-  { display: '99205 New Patient - E/M Level 5', code: '99205' },
-  { display: '99212 Established Patient - E/M Level 2', code: '99212' },
-  { display: '99213 Established Patient - E/M Level 3', code: '99213' },
-  { display: '99214 Established Patient - E/M Level 4', code: '99214' },
-  { display: '99215 Established Patient - E/M Level 5', code: '99215' },
-];
